internal/controllers/gateway/backendtlspolicy_utils.go
package gateway

import (
	"context"

	"github.com/samber/lo"
	apierrors "k8s.io/apimachinery/pkg/api/errors"
	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
	"sigs.k8s.io/controller-runtime/pkg/client"

	"github.com/kong/kubernetes-ingress-controller/v3/internal/gatewayapi"
)

// getBackendTLSPoliciesByHTTPRoute returns the list of BackendTLSPolicies that targets a service
// used as backend by the given HTTPRoute.
func (r *BackendTLSPolicyReconciler) getBackendTLSPoliciesByHTTPRoute(ctx context.Context, httpRoute gatewayapi.HTTPRoute) ([]gatewayapi.BackendTLSPolicy, error) {
	objects := []gatewayapi.BackendTLSPolicy{}
	for _, rule := range httpRoute.Spec.Rules {
		for _, backend := range rule.BackendRefs {
			// no need to check group and kind nilness, as they have a default value in case not specified
			if !((*backend.Group == "" || *backend.Group == "core") && *backend.Kind == "Service") {
				continue
			}

			namespace := httpRoute.Namespace
			if backend.Namespace != nil {
				namespace = string(*backend.Namespace)
			}
			policies := &gatewayapi.BackendTLSPolicyList{}
			if err := r.List(ctx, policies,
				client.InNamespace(namespace),
				client.MatchingFields{backendTLSPolicyTargetRefIndexKey: string(backend.Name)},
			); err != nil {
				return nil, err
			}
			objects = append(objects, policies.Items...)
		}
	}
	return objects, nil
}

// getBackendTLSPolicyAncestors returns the list of Gateways associated to the given BackendTLSPolicy.
// To retrieve such a list, the following steps are performed:
// 1. Get all the the HTTPRoutes that reference the backends targeted by the policy;
// 4. Find all the parents in the HTTPRoute status that have already properly resolved by KIC and have the resolvedRefs condition set to true;
// 3. Return all the successfully resolved Gateways.
func (r *BackendTLSPolicyReconciler) getBackendTLSPolicyAncestors(ctx context.Context, policy gatewayapi.BackendTLSPolicy) ([]gatewayapi.Gateway, error) {
	gateways := []gatewayapi.Gateway{}
	for _, targetRef := range policy.Spec.TargetRefs {
		if (targetRef.Group != "core" && targetRef.Group != "") && targetRef.Kind != "Service" {
			continue
		}

		httpRoutes := gatewayapi.HTTPRouteList{}
		if err := r.Client.List(ctx, &httpRoutes,
			client.MatchingFields{httpRouteBackendRefIndexKey: policy.Namespace + "/" + string(targetRef.Name)},
		); err != nil {
			return nil, err
		}

		for _, httpRoute := range httpRoutes.Items {
			for _, parentRef := range httpRoute.Spec.ParentRefs {
				if *parentRef.Group != gatewayapi.V1Group || *parentRef.Kind != "Gateway" {
					continue
				}

				namespace := httpRoute.Namespace
				if parentRef.Namespace != nil {
					namespace = string(*parentRef.Namespace)
				}

				var resolvedRefsStatus bool
				// Check the resolvedRefs condition is set to true to ensure that all the references are properly resolved
				// and granted by ReferenceGrants.
				for _, parentStatus := range httpRoute.Status.Parents {
					if parentStatus.ControllerName == GetControllerName() &&
						*parentStatus.ParentRef.Group == *parentRef.Group &&
						*parentStatus.ParentRef.Kind == *parentRef.Kind &&
						parentStatus.ParentRef.Name == parentRef.Name &&
						*parentStatus.ParentRef.Namespace == gatewayapi.Namespace(namespace) {
						if _, found := lo.Find(parentStatus.Conditions, func(c metav1.Condition) bool {
							return c.Type == string(gatewayapi.RouteConditionResolvedRefs) && c.Status == metav1.ConditionTrue
						}); found {
							resolvedRefsStatus = true
							break
						}
					}
				}
				if resolvedRefsStatus {
					gateway := &gatewayapi.Gateway{}
					if err := r.Get(ctx, client.ObjectKey{Namespace: namespace, Name: string(parentRef.Name)}, gateway); err != nil {
						// In case the object is not found, we don't want to return an error, but we want to continue.
						if apierrors.IsNotFound(err) {
							continue
						}
						return nil, err
					}
					gateways = append(gateways, *gateway)
				}
			}
		}
	}

	return gateways, nil
}

// setPolicyStatus enforces an ancestorStatus for each Gateway associated to the given policy.
// TODO: Conditions to the policy still to be implemented.
func (r *BackendTLSPolicyReconciler) setPolicyStatus(ctx context.Context, policy gatewayapi.BackendTLSPolicy, gateways []gatewayapi.Gateway) error {
	ancestors := []gatewayapi.PolicyAncestorStatus{}

	// First copy all the ancestorstatuses managed by other controllers.
	for _, ancestor := range policy.Status.Ancestors {
		if ancestor.ControllerName == GetControllerName() {
			continue
		}
		ancestors = append(ancestors, ancestor)
	}

	// Then enforces all the ancestorsStatuses for the Gateways managed by this controller.
	for _, gateway := range gateways {
		ancestor := gatewayapi.PolicyAncestorStatus{
			AncestorRef: gatewayapi.ParentReference{
				Group:     lo.ToPtr(gatewayapi.V1Group),
				Kind:      lo.ToPtr(gatewayapi.Kind("Gateway")),
				Name:      gatewayapi.ObjectName(gateway.Name),
				Namespace: lo.ToPtr(gatewayapi.Namespace(gateway.Namespace)),
			},
			ControllerName: GetControllerName(),
		}
		ancestors = append(ancestors, ancestor)
	}

	newPolicy := policy.DeepCopy()
	newPolicy.Status.Ancestors = ancestors

	return r.Status().Patch(ctx, newPolicy, client.MergeFrom(&policy))
}
